+++ datafusion/expr/src/type_coercion.rs
@@ -57,6 +57,7 @@ pub fn is_numeric(dt: &DataType) -> bool {
         )
 }
 
+pub mod aggregates;
 pub mod binary;
 pub mod functions;
 pub mod other;
